backend/requirements.txt
asgiref==3.9.1
atomicwrites==1.4.1
attrs==25.3.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==45.0.6
defusedxml==0.7.1
Django==3.2.3
django-admin-autocomplete-filter==0.7.1
django-cors-headers==3.13.0
django-filter==21.1
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-extra-fields==3.7.0
filetype==1.2.0
gunicorn==23.0.0
idna==3.10
iniconfig==2.1.0
isort==5.13.2
itypes==1.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mccabe==0.7.0
numpy==2.3.2
oauthlib==3.3.1
packaging==25.0
Pillow==10.0.0
pluggy==0.13.1
psycopg2-binary==2.9.6
py==1.11.0
pycodestyle==2.10.0
pycparser==2.22
pyflakes==3.0.1
PyJWT==2.10.1
pytest==6.2.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python3-openid==3.2.0
pytz==2025.2
requests==2.32.4
requests-oauthlib==2.0.0
six==1.17.0
social-auth-app-django==4.0.0
social-auth-core==4.7.0
sqlparse==0.5.3
toml==0.10.2
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
